tidy

+++ tasks/konstantinov_s_elem_vec_sign_change_count/tests/testgen.h
@@ -8,8 +8,8 @@
 namespace konstantinov_s_elem_vec_sign_change_count {
 // СТАБИЛЬНЫЙ генератор тестовых данных
 // Нужно подать массив с известным колвом смен знаков и одинаковыми знаками первого и последнего элемента
-inline int generateTestData(const EType *examplearr, int arrsz, int sign_change_count, InType &v) {
-  int m = v.size();
+inline int generateTestData(const EType *examplearr, size_t arrsz, int sign_change_count, InType &v) {
+  size_t m = v.size();
   // std::cout<<"Generating "<<m<<"\n";
   if (arrsz == 0 || m == 0) {
     return 0;
@@ -18,7 +18,7 @@ inline int generateTestData(const EType *examplearr, int arrsz, int sign_change_
   if (m < arrsz)  // для особо маленьких размеров
   {
     int res = 0;
-    for (int i = 0; i < m - 1; i++) {
+    for (size_t i = 0; i < m - 1; i++) {
       v[i] = examplearr[i];
       res += (examplearr[i] > 0) != (examplearr[i + 1] > 0);
     }
@@ -26,14 +26,14 @@ inline int generateTestData(const EType *examplearr, int arrsz, int sign_change_
     return res;
   }
 
-  int fullBlocks = m / arrsz;
-  int tail = m % arrsz;
+  size_t fullBlocks = m / arrsz;
+  size_t tail = m % arrsz;
 
   EType *dst = v.data();
 
   // Копируем первую копию массива
   memcpy(dst, examplearr, arrsz * sizeof(EType));
-  int filled = 1;  // число заполненных блоков
+  size_t filled = 1;  // число заполненных блоков
 
   // Удвоительное копирование блоками
   // пока удвоение не превысит количество нужных полных блоков
@@ -51,7 +51,7 @@ inline int generateTestData(const EType *examplearr, int arrsz, int sign_change_
   // Хвост = последний элемент чтобы не считать смены знаков там
   if (tail > 0) {
     EType last = examplearr[arrsz - 1];
-    for (int i = 0; i < tail; ++i) {
+    for (size_t i = 0; i < tail; ++i) {
       dst[fullBlocks * arrsz + i] = last;
     }
   }
